requirements.txt
flask
flask-cors
gunicorn
opencv-python
numpy
pillow
requests
tensorflow
mediapipe
